% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aggregated_covid_race_df}
\alias{aggregated_covid_race_df}
\title{"Aggregated COVID Race Data"}
\format{
A data frame with 7 rows and 10 variables:
\describe{
\item{Race}{Race (or ethnicity in case of Hispanic)}
\item{Cases_Total}{Total number of COVID-19 cases by race or ethnicity}
\item{Deaths_Total}{Total number of COVID-19 deaths by race or ethnicity}
\item{Population_Total}{Total US population by race or ethnicity}
\item{Cases_Per_100K}{Total cases (per 100,000 people) by race or ethnicity}
\item{Deaths_Per_100K}{Total deaths (per 100,000 people) by race or ethnicity}
\item{Population_In_100K}{Total US population (in 100,00's of people) by race or ethnicity}
\item{Cases_Percent}{Percent of COVID-19 cases by race or ethnicity}
\item{Deaths_Percent}{Percent of COVID-19 deaths by race or ethnicity}
\item{Population_Percent}{Percent of US population by race or ethnicity}
}
}
\source{
\url{https://covidtracking.com/race/about#download-the-data}
}
\usage{
aggregated_covid_race_df
}
\description{
A data set containing the number of cases, deaths, and population (in raw
numbers as well as by percentage) by race or ethnicity in the US.
}
\keyword{datasets}
