% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{CovidRaceData}
\alias{CovidRaceData}
\alias{CovidRaceData-package}
\title{COVID-19 Data By Race and Ethnicity}
\description{
This package contains data sets on the COVID-19 pandemic in the US by race and ethnicity.
}
